Updates SectionHeading component
